fix edits from david feedback
