+++ app/components/home-page/landing.tsx
@@ -10,8 +10,6 @@ interface MissionStatementProps {
     images
   }: MissionStatementProps) => {
     const c4cPurple = '#605ACD';
-    const c4cLightPurple = '#F8EDFF';
-    const c4cGray = '#4A4A51';
     return (
       <div className="flex flex-col mb-20">
         <p className="text-4xl font-bold mb-2">Code4Community designs software solutions for</p>
@@ -22,7 +20,7 @@ interface MissionStatementProps {
 
         <div className="grid grid-cols-4 gap-4 content-start">
             {images?.map((image, i) => (
-                <img src={image} alt="" width={'264px'} height={'264px'}/>
+                <img src={image} alt="" key={i} width={'264px'} height={'264px'}/>
             ))}
         </div>
       </div>
