+++ app/routes/_index.tsx
@@ -1,6 +1,6 @@
 import { type MetaFunction } from "@remix-run/node";
-import WeAreBlock from "~/components/info";
-import MissionStatement from "~/components/landing";
+import WeAreBlock from "~/components/home-page/we-are-block";
+import MissionStatement from "~/components/home-page/landing";
 
 export const meta: MetaFunction = () => {
   return [
@@ -11,8 +11,6 @@ export const meta: MetaFunction = () => {
 
 export default function Index() {
   const c4cPurple = '#605ACD';
-  const c4cLightPurple = '#F8EDFF';
-  const c4cGray = '#4A4A51';
 
   const weLove = ['people and causes we love.'];
   const images = ['https://images.unsplash.com/photo-1499796683658-b659bc751db1?q=80&w=1974&auto=format&fit=crop&ixlib=rb-4.0.3&ixid=M3wxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8fA%3D%3D', 'https://images.unsplash.com/photo-1499796683658-b659bc751db1?q=80&w=1974&auto=format&fit=crop&ixlib=rb-4.0.3&ixid=M3wxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8fA%3D%3D', 'https://images.unsplash.com/photo-1499796683658-b659bc751db1?q=80&w=1974&auto=format&fit=crop&ixlib=rb-4.0.3&ixid=M3wxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8fA%3D%3D', 'https://images.unsplash.com/photo-1499796683658-b659bc751db1?q=80&w=1974&auto=format&fit=crop&ixlib=rb-4.0.3&ixid=M3wxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8fA%3D%3D'];
